+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Fixed
+- Fixed step param value for fields in category edit
+
 ## [1.1.2]
 ### Fixed
 - Fixed offer price getting, when got unknown value from the settings